+++ src/test/java/nextstep/subway/ui/acceptance/AcceptanceTest.java
@@ -1,4 +1,4 @@
-package nextstep.subway.acceptance;
+package nextstep.subway.ui.acceptance;
 
 import nextstep.subway.utils.DatabaseCleanup;
 import org.junit.jupiter.api.BeforeEach;
