+++ src/test/java/nextstep/subway/ui/acceptance/StationSteps.java
@@ -1,4 +1,4 @@
-package nextstep.subway.acceptance;
+package nextstep.subway.ui.acceptance;
 
 import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
